server/pyproject.toml
[project]
name = "console-mcp-server"
version = "0.1.0"
description = "Prototype API service for the Console MCP orchestration layer"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.109.0",
  "uvicorn[standard]>=0.25.0",
  "pydantic>=2.5.0",
  "pydantic-settings>=2.0.0",
  "sqlalchemy>=2.0.25",
  "structlog>=23.2.0",
  "GitPython>=3.1.43",
  "httpx>=0.26.0",
  "PyYAML>=6.0.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.26.0",
  "pytest>=8.2.0",
]

[project.scripts]
console-mcp-server = "console_mcp_server.main:run"
console-mcp-server-dev = "console_mcp_server.main:run_dev"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["src"]
